Add auth tests
This configures permissions so that it sends more accurate error
messages when not in production. Essentially, graphql-shield will
always send a not authorized error, unless you are in debug mode.
This is a security feature, but in test we want to make sure that
the errors are being thrown for the right reasons.