Add layout to frontend and search update to query
This add components for layout on the frontend of the site,
including a BookCard component, Meta component and a Page
component. It also adds a _document page that's needed for
styled-components.

On the backend, it updates the searchBook mutation to be a
query. Now that the call isn't adding books, it makes more sense
for it to be a query instead of a mutation.