Add tests for BookGallery and Signout
This adds tests for the BookGallery component. It adds a testid to
the component itself, and then tests for the happy and error paths.

This adds tests for the Signout component. It adds callbacks for
the toast calls to handle the error state in the component itself,
and it also removes the Link from the component. It seems like
routing in this way isn't recommended (See
https://github.com/vercel/next.js/issues/5205 ) so we'll use it
this way until a more official mock is released. It adds tests
for the happy and error paths.