Add search functionality and refactor BookGallery
This adds search functionality and a new SearchBar component that
can be placed anywhere in the app to execute searches. It leaves the
handling up to another page or component via a handleSearch prop.
It also adds a search page that displays results.

This also updates the BookGallery component to just display books.
There is a new BookCategorySearch component that will take the
functionality of querying that the BookGallery did before. This
makes more sense from a design perspective because it's more
common to just want to display books than query via a category