Add select bookshelf via buttongroup dropdown
This adds the ability to select which bookshelf the user wants
to add a book to to the AddToBookshelf component. It also adds
some buttongroup styling to make that work. I've also worked
a bit on the accessibility of this component. The user can
select with the enter key, and navigate with the arrow keys.
There's still more work to be done here though.