Refactor to use Google Books ID as primary key on book
This is a fairly major refactor to use the Google Books ID as the
primary key. It updates all dof the tests to use googleBooksId. I
decided to keep it as the googleBooksId instead of just id because
it adds clarity to the application on what that piece of data is.