Add reading progress styling and minor improvements
This adds styling to the reading progress on the Book page. It adds a time
to go display, and a little bar graph of how far into the book the user is.

It also makes some style tweaks to the Book page.

It adds an outside click handler to the AddToBookshelf component which
closes the dropdown when the click is outside the element.