Add query to URL on search page
This adds the search term to the URL so that go back functionality works
as expected. Ie., the search is still displayed as it was the last time
the user was on the page. This also means updating the caching policy on
the search to cache-first. That should be fine, give that the searches
won't change from session to session since there's no way for the user
to update what gets returned from the search.

Resolves #2