Close AddToBookshelf on select and reset after adding
This updates the AddToBookshelf component so that it closes after selecting
a shelf from the dropdown. It also allows the user to click on a dropdown
item again to deselect. Finally, it resets the dropdown after adding adding
an item to a shelf, which allows for better interaction when adding to
multiple shelves.

Resolves #4 #5