Add start and end date and status to goals
This adds a start date and an end date to the backend model for goals. It defines the fields in
the model and the schema, as well as adds a database migration.

By default, the status for a book will be not started and the user wil be able to define what
status the book is in for their reading history. This will also allow for a 'currently reading'
page on the site without having to tie it to bookshelves, ala Goodreads.

This update also fixes a few broken tests