Fix handling of local dates on the frontend
There was a bug with UTC/local dates where the frontend was sending dates in local time
but they were being interpreted by the backend as UTC time, which was leading to incorrect
dates being displayed.

This adds a function to manually parse date strings into ISO local time, which sets them
correctly in the database. It also updates the display for formatDate to use the parsing
if a string is passed in.