Update ReadingCard to display time
This adds the time that the reading was added to the ReadingCard component. It also
updates the formatDate file to add a formatTime function. Finally, it moves the page
count to the top of the ReadingCard rather than the date. Since this is a list of
readings, it makes more sense to have the page # be the prominent piece of info