Add currently reading page and type resolver for goals
This adds the beginnings of the currently reading page, including listing goals and the
books that they apply to. For now, this page will just list the book goals, but bookshelf
goals could be incorporated after.

It also adds a type resolver to the goals schema sto just fetch goals that apply to a Book
or BookShelf type.