Add privacy level to all user created data
This adds a privacy table and associated privacy levels to the database in a migration.
It also adds privacy level columns to goals, bookshelves, notes, and readings.

This is in preparation for adding friends to the application. Next steps for this
will be to add it into the schema and resolvers for creating and updating on the
backend. And then of course, doing the work of adding it on the frontend.