Update bookshelf to include setting privacy
This updates the create and update bookshelf components to include the ability to set
the privacy level of the shelf when creating or updating.

It also refactors the form for creation/update into a single form that is implemented
by Create and Update. Finally, it also shifts the Create and Update components into
folders within the Bookshelf folder.